Refactor the War team-selection page into subcomponents for the char grid and mow grid.
